move: Vote 폴더 구조 변경 (meeting.vote -> vote)
