+++ src/main/java/com/grepp/synapse4/app/model/notification/NotificationService.java
@@ -1,7 +1,7 @@
 package com.grepp.synapse4.app.model.notification;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.grepp.synapse4.app.model.meeting.entity.vote.VoteMember;
+import com.grepp.synapse4.app.model.vote.entity.VoteMember;
 import com.grepp.synapse4.app.model.notification.code.NotificationType;
 import com.grepp.synapse4.app.model.notification.code.NotificationEventInfo;
 import com.grepp.synapse4.app.model.notification.dto.NotificationDto;
