+++ src/main/java/com/grepp/synapse4/app/model/vote/repository/VoteRepository.java
@@ -1,6 +1,6 @@
-package com.grepp.synapse4.app.model.meeting.repository.vote;
+package com.grepp.synapse4.app.model.vote.repository;
 
-import com.grepp.synapse4.app.model.meeting.entity.vote.Vote;
+import com.grepp.synapse4.app.model.vote.entity.Vote;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
